Reader fixes

+++ src/dvpio/read/image/openslide.py
@@ -101,7 +101,7 @@ def read_openslide(path: str, chunk_size: tuple[int, int] = (10000, 10000), pyra
     chunk_coords = _create_tiles(dimensions=dimensions, tile_size=chunk_size, min_coordinates=(0, 0))
 
     # Load chunkwise (parallelized with dask.delayed)
-    chunks = _chunk_factory(_get_img, slide=slide, coords=chunk_coords, n_channel=4, level=0)
+    chunks = _chunk_factory(_get_img, slide=slide, coords=chunk_coords, n_channel=4, dtype=np.uint8, level=0)
 
     # Assemble into a single dask array
     array = _assemble(chunks)
